Add SextantSample.Avalonia.sln
